requirements.txt
langchain-openai>=0.1.0
langchain>=0.1.16,<0.2
langchain-core>=0.1.16,<0.2
langchainhub
langchain-community>=0.0.17,<0.1
langsmith>=0.0.40
python-dotenv
streamlit
streamlit_lottie
fastapi
uvicorn
langserve
sse_starlette
bs4
beautifulsoup4
pypdf
PyPDF2
chromadb
faiss-cpu
groq
cassio
wikipedia
arxiv
sentence_transformers
pydantic<2.0
SpeechRecognition
gTTS
pyaudio
pipwin
huggingface_hub
transformers
